Parcel as interface #1053
