+++ commercetools-models/src/main/java/io/sphere/sdk/orders/Parcel.java
@@ -1,48 +1,23 @@
 package io.sphere.sdk.orders;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
 import io.sphere.sdk.models.CreationTimestamped;
 
 import javax.annotation.Nullable;
 import java.time.ZonedDateTime;
 
-public final class Parcel extends Base implements CreationTimestamped {
-    private final String id;
-    private final ZonedDateTime createdAt;
-    @Nullable
-    private final ParcelMeasurements measurements;
-    @Nullable
-    private final TrackingData trackingData;
-
-    @JsonCreator
-    private Parcel(final String id, final ZonedDateTime createdAt, @Nullable final ParcelMeasurements measurements, @Nullable final TrackingData trackingData) {
-        this.id = id;
-        this.createdAt = createdAt;
-        this.measurements = measurements;
-        this.trackingData = trackingData;
+public interface Parcel extends CreationTimestamped {
+    static Parcel of(final String id, final ZonedDateTime createdAt, @Nullable final ParcelMeasurements measurements, @Nullable final TrackingData trackingData) {
+        return new ParcelImpl(id, createdAt, measurements, trackingData);
     }
 
-    public static Parcel of(final String id, final ZonedDateTime createdAt, @Nullable final ParcelMeasurements measurements, @Nullable final TrackingData trackingData) {
-        return new Parcel(id, createdAt, measurements, trackingData);
-    }
-
-    public String getId() {
-        return id;
-    }
+    String getId();
 
     @Override
-    public ZonedDateTime getCreatedAt() {
-        return createdAt;
-    }
+    ZonedDateTime getCreatedAt();
 
     @Nullable
-    public ParcelMeasurements getMeasurements() {
-        return measurements;
-    }
+    ParcelMeasurements getMeasurements();
 
     @Nullable
-    public TrackingData getTrackingData() {
-        return trackingData;
-    }
+    TrackingData getTrackingData();
 }
